pyproject.toml
[project]
name = "lyrics-client"
version = "0.1.0"
description = "MCP Client for Lyric interpretation app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ipykernel>=6.30.0",
    "mcp>=1.11.0",
    "openai>=1.96.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "webdriver-manager>=4.0.2",
]
